requirements.txt
eth-account>=0.5.0
web3>=5.0.0
requests>=2.25.0
websocket-client>=1.0.0
python-dotenv>=0.19.0